use builder pattern for running genesis simulation
