+++ iris-mpc-upgrade-hawk/tests/workflows/genesis_103/runner.rs
@@ -5,6 +5,7 @@ use crate::utils::{
     irises,
     modifications::{ModificationInput, ModificationType},
     mpc_node::{MpcNode, MpcNodes},
+    plaintext_genesis::PlaintextGenesis,
     resources::{self},
     s3_deletions::{get_aws_clients, upload_iris_deletions},
     HawkConfigs, IrisCodePair, TestError, TestRun, TestRunContextInfo,
@@ -89,7 +90,8 @@ impl TestRun for Test {
         let config = &self.configs[0];
         let plaintext_irises = get_irises();
         let expected = Arc::new(
-            MpcNode::simulate_genesis(DEFAULT_GENESIS_ARGS, config, &plaintext_irises, vec![])
+            PlaintextGenesis::new(DEFAULT_GENESIS_ARGS, config, &plaintext_irises)
+                .run()
                 .await
                 .unwrap(),
         );
