+++ iris-mpc-upgrade-hawk/tests/workflows/genesis_104/runner.rs
@@ -129,7 +129,8 @@ impl TestRun for Test {
         let plaintext_irises = get_irises();
         todo!("add modifications to the genesis simulation");
         let expected = Arc::new(
-            MpcNode::simulate_genesis(DEFAULT_GENESIS_ARGS, config, &plaintext_irises, vec![])
+            PlaintextGenesis::new(DEFAULT_GENESIS_ARGS, config, &plaintext_irises)
+                .run()
                 .await
                 .unwrap(),
         );
